+++ src/App.tsx
@@ -1,15 +1,12 @@
 import React from 'react';
 import './App.css';
 import './index.css';
-import AppRouter from './utils/AppRouter';
-import BaseWrapper from './utils/base-wrapper';
+import AppRouter from './utils/app-router';
 
 const App = () => {
   return (
-    <div className="App">
-      <BaseWrapper>
-        <AppRouter/>
-      </BaseWrapper>
+    <div className="App p-4 border rounded-md shadow-md text-center text-text">
+      <AppRouter/>
     </div>
   );
 };