+++ src/pages/home.tsx
@@ -4,6 +4,7 @@ import { Container, Typography } from '@mui/material';
 import { fetchPopularMovies } from '../endpoints/movie';
 import MovieListContainer from '../components/features/movie-list/movie-list-container';
 import SearchForm from '../components/features/search/search-form';
+import BaseWrapper from '../utils/base-wrapper';
 
 const HomePage: React.FC = () => {
   const [movies, setMovies] = useState<Movie[]>([]);
@@ -29,13 +30,15 @@ const HomePage: React.FC = () => {
   if (error) return <Typography color="error" textAlign="center">{error}</Typography>;
 
   return (
-    <Container>
-      <SearchForm movies={movies} setMovies={setMovies} setLoading={setLoading} setHeaderLabel={setHeaderLabel}/>
-      <Typography variant="h4" textAlign="center" my={3}>
-        {headerLabel}
-      </Typography>
-      <MovieListContainer movies={movies} loading={loading}/>
-    </Container>
+    <BaseWrapper>
+      <Container>
+        <SearchForm movies={movies} setMovies={setMovies} setLoading={setLoading} setHeaderLabel={setHeaderLabel}/>
+        <Typography variant="h4" textAlign="center" my={3}>
+          {headerLabel}
+        </Typography>
+        <MovieListContainer movies={movies} loading={loading}/>
+      </Container>
+    </BaseWrapper>
   );
 };