+++ src/pages/movie.tsx
@@ -1,8 +1,11 @@
 import MovieDetailsContainer from '../components/features/movie-details/movie-details-container';
+import BaseWrapper from '../utils/base-wrapper';
 
 const MoviePage = () => {
   return (
-    <MovieDetailsContainer/>
+    <BaseWrapper>
+      <MovieDetailsContainer/>
+    </BaseWrapper>
   );
 };